refactor: move signIn and signUp functions to separate file and remove unused imports
