+++ app/login/page.tsx
@@ -1,61 +1,12 @@
 import Link from 'next/link'
-import { headers, cookies } from 'next/headers'
-import { createClient } from '@/utils/supabase/server'
-import { redirect } from 'next/navigation'
+import { signIn, signUp } from './actions'
 
 export default function Login({
   searchParams,
 }: {
   searchParams: { message: string }
 }) {
-  const signIn = async (formData: FormData) => {
-    'use server'
-
-    const email = formData.get('email') as string
-    const password = formData.get('password') as string
-    const cookieStore = cookies()
-    const supabase = createClient(cookieStore)
-
-    const { error } = await supabase.auth.signInWithPassword({
-      email,
-      password,
-    })
-
-    if (error) {
-      return redirect('/login?message=Could not authenticate user')
-    }
-
-    return redirect('/')
-  }
-
-  const signUp = async (formData: FormData) => {
-    'use server'
-
-    const origin = headers().get('origin')
-    const email = formData.get('email') as string
-    const password = formData.get('password') as string
-    const cookieStore = cookies()
-    const supabase = createClient(cookieStore)
-
-    const { error } = await supabase.auth.signUp({
-      email: email,
-      password: password,
-          });
-          if (error) {
-              console.log(error);
-              if (error == null){
-                console.log(error);
-                  
-              } else {
-                  console.log(error);
-                  // Password must be 6 letters
-                  return redirect("/login?message=There was an error while creating your account. Please try again.")
-
-              }
-          } else{
-            return redirect('/login?message=Check email to continue sign in process')
-          }
-  }
+  
 
   return (
     <div className="flex-1 flex flex-col w-full px-8 sm:max-w-md justify-center gap-2">
